requirements.txt
# Service agent Gifters : SmolAgent + MCP + Mistral-7B
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
python-dotenv>=1.0.0

# SmolAgent avec MCP (streamable-http) et Transformers (Mistral-7B-Instruct-v0.3)
smolagents[mcp,transformers]>=1.24.0
# torch / transformers sont installés via smolagents[transformers]
